+++ public/images/dataSets/sharks.json
@@ -1,32 +1,32 @@
 [
-{"image": "images/dataSets/sharks/carcharhinus_leucas/VP UF122794 lingual copy.png", "category": "carcharhinus_leucas"},
 {"image": "images/dataSets/sharks/carcharhinus_leucas/VP UF122910B lingual copy.png", "category": "carcharhinus_leucas"},
+{"image": "images/dataSets/sharks/carcharhinus_leucas/VP UF122794 lingual copy.png", "category": "carcharhinus_leucas"},
 {"image": "images/dataSets/sharks/carcharhinus_leucas/VP UF122563E lingual copy.png", "category": "carcharhinus_leucas"},
 {"image": "images/dataSets/sharks/carcharhinus_leucas/VP UF122563B lingual copy.png", "category": "carcharhinus_leucas"},
 {"image": "images/dataSets/sharks/carcharhinus_leucas/VP UF122910A lingual copy.png", "category": "carcharhinus_leucas"},
-{"image": "images/dataSets/sharks/carcharodon_carcharias/VP UF-TRO2369 lingual copy.png", "category": "carcharodon_carcharias"},
-{"image": "images/dataSets/sharks/carcharodon_carcharias/VP UF-TRO2371 lingual copy.png", "category": "carcharodon_carcharias"},
 {"image": "images/dataSets/sharks/carcharodon_carcharias/VP UF-TRO2366 lingual copy.png", "category": "carcharodon_carcharias"},
+{"image": "images/dataSets/sharks/carcharodon_carcharias/VP UF-TRO2371 lingual copy.png", "category": "carcharodon_carcharias"},
+{"image": "images/dataSets/sharks/carcharodon_carcharias/VP UF-TRO2369 lingual copy.png", "category": "carcharodon_carcharias"},
+{"image": "images/dataSets/sharks/carcharodon_carcharias/VP UF3655 lingual copy.png", "category": "carcharodon_carcharias"},
 {"image": "images/dataSets/sharks/carcharodon_carcharias/VP UF-TRO2367 lingual copy.png", "category": "carcharodon_carcharias"},
-{"image": "images/dataSets/sharks/carcharodon_carcharias/VP UF-TRO2368 lingual copy.png", "category": "carcharodon_carcharias"},
-{"image": "images/dataSets/sharks/galeocerdo_cuvier/VP UF-TRO8951 lingual copy.png", "category": "galeocerdo_cuvier"},
-{"image": "images/dataSets/sharks/galeocerdo_cuvier/VP UF-TRO8941 lingual copy.png", "category": "galeocerdo_cuvier"},
-{"image": "images/dataSets/sharks/galeocerdo_cuvier/VP UF-TRO8939 lingual copy.png", "category": "galeocerdo_cuvier"},
 {"image": "images/dataSets/sharks/galeocerdo_cuvier/VP UF-TRO8948 lingual copy.png", "category": "galeocerdo_cuvier"},
 {"image": "images/dataSets/sharks/galeocerdo_cuvier/VP UF-TRO8943 lingual copy.png", "category": "galeocerdo_cuvier"},
-{"image": "images/dataSets/sharks/hemipristis_serra/VP UF-TRO 7579  lingual copy.png", "category": "hemipristis_serra"},
+{"image": "images/dataSets/sharks/galeocerdo_cuvier/VP UF-TRO8941 lingual copy.png", "category": "galeocerdo_cuvier"},
+{"image": "images/dataSets/sharks/galeocerdo_cuvier/VP UF-TRO8939 lingual copy.png", "category": "galeocerdo_cuvier"},
+{"image": "images/dataSets/sharks/galeocerdo_cuvier/VP UF-TRO8045 lingual copy.png", "category": "galeocerdo_cuvier"},
 {"image": "images/dataSets/sharks/hemipristis_serra/VP UF-TRO 7575 lingual copy.png", "category": "hemipristis_serra"},
 {"image": "images/dataSets/sharks/hemipristis_serra/VP UF-TRO 7580 lingual copy.png", "category": "hemipristis_serra"},
-{"image": "images/dataSets/sharks/hemipristis_serra/VP UF-TRO 7576 lingual copy.png", "category": "hemipristis_serra"},
+{"image": "images/dataSets/sharks/hemipristis_serra/VP UF-TRO 7540 lingual copy.png", "category": "hemipristis_serra"},
+{"image": "images/dataSets/sharks/hemipristis_serra/VP UF-TRO 7579  lingual copy.png", "category": "hemipristis_serra"},
 {"image": "images/dataSets/sharks/hemipristis_serra/VP UF 229824 lingual copy.png", "category": "hemipristis_serra"},
-{"image": "images/dataSets/sharks/isurus_hastalis/VP UF-TRO7127 lingual copy.png", "category": "isurus_hastalis"},
 {"image": "images/dataSets/sharks/isurus_hastalis/VP UF-TRO7125 lingual copy.png", "category": "isurus_hastalis"},
-{"image": "images/dataSets/sharks/isurus_hastalis/VP UF-TRO5637 lingual copy.png", "category": "isurus_hastalis"},
-{"image": "images/dataSets/sharks/isurus_hastalis/VP UF-TRO7128 lingual copy.png", "category": "isurus_hastalis"},
 {"image": "images/dataSets/sharks/isurus_hastalis/VP UF-TRO3913 lingual copy.png", "category": "isurus_hastalis"},
+{"image": "images/dataSets/sharks/isurus_hastalis/VP UF-TRO7127 lingual copy.png", "category": "isurus_hastalis"},
+{"image": "images/dataSets/sharks/isurus_hastalis/VP UF-TRO7128 lingual copy.png", "category": "isurus_hastalis"},
+{"image": "images/dataSets/sharks/isurus_hastalis/VP UF-TRO7129 lingual copy.png", "category": "isurus_hastalis"},
+{"image": "images/dataSets/sharks/otodus_megalodon/VP UF-TRO2755 lingual copy.png", "category": "otodus_megalodon"},
 {"image": "images/dataSets/sharks/otodus_megalodon/VP UF-TRO4484 lingual copy.png", "category": "otodus_megalodon"},
-{"image": "images/dataSets/sharks/otodus_megalodon/VP UF-TRO2735 lingual copy.png", "category": "otodus_megalodon"},
 {"image": "images/dataSets/sharks/otodus_megalodon/VP UF-TRO4486 lingual copy.png", "category": "otodus_megalodon"},
 {"image": "images/dataSets/sharks/otodus_megalodon/VP GS1381 lingual copy.png", "category": "otodus_megalodon"},
-{"image": "images/dataSets/sharks/otodus_megalodon/VP UF-TRO4485 lingual copy.png", "category": "otodus_megalodon"}
+{"image": "images/dataSets/sharks/otodus_megalodon/VP UF-TRO2735 lingual copy.png", "category": "otodus_megalodon"}
 ]
