+++ public/index.html
@@ -105,7 +105,7 @@
 <script src="https://cdn.jsdelivr.net/npm/@tensorflow/tfjs"> </script>
 <script src="https://cdn.jsdelivr.net/npm/@tensorflow/tfjs-vis"></script>
 <!-- script.js- logic for tab navigation -->
-<script src="./fossilML.js"></script>
+<script type="module" src="./fossilML.js"></script>
 
 
 </body>
